Move Brand to components, named export example
