+++ src/components/BarSelectionBrands/index.js
@@ -1,0 +1,1 @@
+export { default as BarSelectionBrands } from './BarSelectionBrands';
